Add Support for More GLM Link Functions
Add a function that maps most all GLM family functions in the glm() function in R to an appropriate specification in PFA for generating documents

Add a pfa() function that creates a complete pfa document from an object and an S3 method for glm models to be generated via pfa(). This decreases barriers to model creation for simpler GLM models.